fix: DraftEditor crashes in Safari after IME composition
When typing using IME at the end of a composition, draft-js will do
a full rerender. Which is done by changing the key of DraftEditor component.
If we change the key of a react component, the component itself
will not remount, but any children inside this component will do.
Therefore, children are unmounted and mounted again, which causes editor
crash issue in Safari. Since React reconciliation fail to diff previous DOM
and current DOM due to some of them are considered as 'not-existing'
which can not be removed by the reconsiliation process.

Since this patch, key of existing block won't change hence its children won't remount.